auth: move out of `data` package to root
Summary:
The `AuthProvider` and `AuthContext` types aren’t intended to be
specific to the `DataProvider` interface, so we move the module to root
level for clarity.

Test Plan:
There aren’t any existing references to this module (either in this repo
or internally to Google), so it suffices that tests pass.

wchargin-branch: auth-to-root
wchargin-source: 8c9fb48229690b843da2e47b061d12c33fb6355c
